Changes to the backend. DO NOT MAKE LIVE
In order to start allowing for more courses and all the user data input
on the website to actually work, a significant about of the codebase has
to be edited first to allow for these changes. In the meantime, this
commit is to be used for development ONLY and should not be made live.